Update UrlHasher._handle_html() to navigate to a local file
When testing against different sites, I found that for more complex sites there
was not always enough time for everything to process and render. Unfortunately
the setContent() coroutine does not allow you to pass additional options. I
have switched to using the goto() coroutine instead, as this allows me ot use
the waitUntil option to give time for the page to process and fully load.