Handle browser timeout while waiting for content
Adjust the browser's timeout while waiting for the events given in the
waitUntil to occur from the default of 30 seconds to five seconds. Wrap
everything in a try/except block and pull whatever content is rendered at the
end of the timeout as a fallback. If a site has long-polling or similar side
activity, then the networkidle2 event may never occur. Handling the timeout
will allow us to retrieve a site that has likely loaded already.